requirements.txt
# Core dependencies
requests
python-dotenv
rich
questionary
curl_cffi
paho-mqtt
